refactor: move filters to core
