+++ backend/src/main.ts
@@ -7,7 +7,7 @@ import { ValidationPipe } from '@nestjs/common';
 import { DataSource } from 'typeorm';
 import { TypeormStore } from 'connect-typeorm';
 import { AppModule } from './app.module';
-import { AxiosExceptionFilter } from './filters';
+import { AxiosExceptionFilter } from './core/filters';
 import { AuthProvider, Session } from './core/entities';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
